refactor : (Service) Domain Layer에서 Application Layer로 이동
