+++ orury-client/src/main/java/org/orury/client/comment/interfaces/message/CommentMessage.java
@@ -1,4 +1,4 @@
-package org.orury.client.comment.application.message;
+package org.orury.client.comment.interfaces.message;
 
 import lombok.Getter;
