+++ orury-client/src/main/java/org/orury/client/gym/application/GymFacade.java
@@ -5,7 +5,6 @@ import org.orury.client.gym.interfaces.response.GymResponse;
 import org.orury.client.gym.interfaces.response.GymReviewStatistics;
 import org.orury.client.gym.interfaces.response.GymsResponse;
 import org.orury.client.review.service.ReviewService;
-import org.orury.domain.gym.domain.GymService;
 import org.orury.domain.gym.domain.dto.GymDto;
 import org.orury.domain.gym.domain.dto.GymLikeDto;
 import org.springframework.stereotype.Service;
