+++ orury-client/src/main/java/org/orury/client/gym/application/GymService.java
@@ -1,4 +1,4 @@
-package org.orury.domain.gym.domain;
+package org.orury.client.gym.application;
 
 import org.orury.domain.gym.domain.dto.GymDto;
 import org.orury.domain.gym.domain.dto.GymLikeDto;
