+++ orury-client/src/main/java/org/orury/client/user/interfaces/UserController.java
@@ -3,6 +3,7 @@ package org.orury.client.user.interfaces;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.orury.client.comment.application.CommentService;
 import org.orury.client.global.WithCursorResponse;
 import org.orury.client.review.service.ReviewService;
 import org.orury.client.user.application.UserFacade;
@@ -13,7 +14,6 @@ import org.orury.client.user.interfaces.response.MyPostResponse;
 import org.orury.client.user.interfaces.response.MyReviewResponse;
 import org.orury.client.user.interfaces.response.MypageResponse;
 import org.orury.domain.base.converter.ApiResponse;
-import org.orury.domain.comment.domain.CommentService;
 import org.orury.domain.comment.domain.dto.CommentDto;
 import org.orury.domain.global.constants.NumberConstants;
 import org.orury.domain.review.dto.ReviewDto;
