+++ orury-client/src/test/java/org/orury/client/config/ControllerTest.java
@@ -7,7 +7,9 @@ import org.orury.client.auth.controller.AuthController;
 import org.orury.client.auth.jwt.JwtTokenProvider;
 import org.orury.client.auth.service.AuthService;
 import org.orury.client.auth.strategy.LoginStrategyManager;
-import org.orury.client.comment.application.CommentController;
+import org.orury.client.comment.application.CommentService;
+import org.orury.client.comment.interfaces.CommentController;
+import org.orury.client.gym.application.GymService;
 import org.orury.client.gym.interfaces.GymController;
 import org.orury.client.post.interfaces.PostController;
 import org.orury.client.review.controller.ReviewController;
@@ -16,11 +18,9 @@ import org.orury.client.review.service.ReviewReactionService;
 import org.orury.client.review.service.ReviewService;
 import org.orury.client.user.interfaces.UserController;
 import org.orury.common.config.SlackMessage;
-import org.orury.domain.comment.domain.CommentService;
-import org.orury.domain.gym.domain.GymService;
 import org.orury.domain.post.domain.PostLikeService;
-import org.orury.domain.user.domain.UserService;
 import org.orury.domain.post.domain.PostService;
+import org.orury.domain.user.domain.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
